fix(catalog-ui): move components into specific folder
