+++ apps/catalog-ui/README.md
@@ -62,12 +62,12 @@ The library organizes components by category. For each component, documentation 
 ### **Navigation Components**
 
 - **NavigationMenu** — Tab-based navigation menu for module pages
-  → Documentation is inside: [`NavigationMenu.md`](./docs/components/NavigationMenu.md)
+  → Documentation is inside: [`NavigationMenu.md`](docs/components/menu/NavigationMenu.md)
 
 ### **Table Components**
 
 - **GenericEntityTable** — Minimal QTable wrapper for simple entity lists
-  → Documentation is inside: [`GenericEntityTable.md`](./docs/components/GenericEntityTable.md)
+  → Documentation is inside: [`GenericEntityTable.md`](docs/components/table/GenericEntityTable.md)
 
 - ✨ **InformationCard** — Reusable card component for displaying labeled read-only information with optional loading states
   → Documentation is inside: [`InformationCard.md`](./docs/components/card/InformationCard.md)
