+++ apps/catalog-ui/docs/components/table/GenericEntityTable.md
